Moved creation of example anchor pane.
